Update
Added method to check if a rack contains duplicate letters. Modified run() so that the score of words with blank tiles in them have the value of the tile the blank changed to be = 0. Added variables String "big" and int "bigS" in order to track the largest value word possible in the rack, and the value of that word